添加时间耗时

+++ marketing-cs/src/main/java/com/br/marketing/entity/MarketingTask.java
@@ -32,8 +32,8 @@ public class MarketingTask {
     private Integer actualNumber;
     private Integer increment;
 
-    private Integer begin;
-    private Integer end;
+    private Long begin;
+    private Long end;
     private String tableName;
 
     /**
